refactor: delete user modal ui
