+++ src/features/remove-user/ui/RemoveUserModal.module.scss
@@ -1,0 +1,26 @@
+.buttonBlock {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+
+  margin-bottom: 1.5rem;
+  padding: 1.5rem;
+
+  button {
+    width: 25%;
+  }
+}
+
+.modalContentBlock {
+  display: flex;
+  flex-direction: column;
+  gap: 1.13rem;
+  align-items: flex-start;
+
+  width: 100%;
+
+  .selectBlock {
+    width: 100%;
+    margin-top: 1.5rem;
+  }
+}
